requirements.txt
# 豆瓣书单同步与 Calibre 集成自动化项目依赖

# 基础依赖
python-dotenv>=0.19.2
pyyaml>=6.0
click>=8.1.3

# 网络请求与解析
requests>=2.28.1
beautifulsoup4>=4.11.1
lxml>=4.9.1

# 数据库
sqlalchemy>=1.4.41
alembic>=1.8.1
psycopg2-binary>=2.9.5  # PostgreSQL 支持（可选）

# 调度系统
apscheduler>=3.9.1

# Z-Library API 客户端
zlibrary>=0.1.0  # Z-Library API 客户端包

# 日志与监控
loguru>=0.6.0

# 测试工具
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# 代码质量工具
flake8>=6.0.0
pylint>=2.15.10
black>=22.12.0
isort>=5.11.4

# 类型检查
mypy>=0.991
types-requests>=2.28.11.5
types-pyyaml>=6.0.12.2

# 工具库
python-dateutil>=2.8.2
fuzzy-match>=0.0.1  # 模糊匹配算法
tqdm>=4.64.1  # 进度条